fix 32bit

+++ base/Base.jl
@@ -425,8 +425,8 @@ include("weakkeydict.jl")
 include("scopedvalues.jl")
 
 # used by task.jl
-const task_times_per_thread = ScopedValues.ScopedValue{Vector{UInt}}()
-const sleep_times_per_thread = ScopedValues.ScopedValue{Vector{UInt}}()
+const task_times_per_thread = ScopedValues.ScopedValue{Vector{UInt64}}() # time_ns always returns UInt64
+const sleep_times_per_thread = ScopedValues.ScopedValue{Vector{UInt64}}()
 
 # metaprogramming
 include("meta.jl")
